discord_push.py
"""
Discord Push Notifications - Sends KOL monitoring updates to Discord via webhooks.
"""
import json
import os
import time
import urllib.request
from datetime import datetime, timedelta

import psycopg2
import psycopg2.extras
from config import DB_DSN, DISCORD_WEBHOOKS

def send_discord(webhook_url, embed):
    if not webhook_url:
        print(f"  ⚠️ Missing webhook for embed, skipping send")
        return False
    
    data = {"username": "KOL Monitor", "avatar_url": "https://i.imgur.com/g91gVAM.png", "embeds": [embed]}
    headers = {'Content-Type': 'application/json', 'User-Agent': 'Mozilla/5.0'}
    
    try:
        req = urllib.request.Request(webhook_url, data=json.dumps(data).encode('utf-8'), headers=headers)
        resp = urllib.request.urlopen(req, timeout=15)
        if not (200 <= resp.status < 300):
            print(f"  ❌ Discord send error: {resp.status} {resp.read().decode()}")
        return 200 <= resp.status < 300
    except Exception as e:
        print(f"  ❌ Discord send error: {e}")
        return False

def get_db():
    return psycopg2.connect(DB_DSN)

def push_new_posts(webhooks, since_minutes):
    db = get_db()
    cur = db.cursor(cursor_factory=psycopg2.extras.RealDictCursor)
    
    cur.execute("""
        SELECT p.*, k.name as kol_name, k.category, 'Twitter' as platform FROM twitter_posts p
        JOIN kols k ON k.id = p.kol_id WHERE p.is_notified = FALSE
        ORDER BY p.captured_at ASC
    """)
    
    posts = cur.fetchall()
    db.close()
    
    if not posts: return
    
    for post in posts:
        color = 0x1DA1F2 
        embed = {
            "title": f"New Tweet from {post['kol_name']}",
            "url": post.get('post_url'),
            "description": (post['content'][:1024] if post['content'] else "No content."),
            "color": color,
            "fields": [
                {"name": "Likes", "value": str(post['likes'] or 0), "inline": True},
                {"name": "Retweets", "value": str(post.get('reposts', 0)), "inline": True},
            ],
            "footer": {"text": f"Category: {post['category'] or 'N/A'}"},
            "timestamp": datetime.now().isoformat() 
        }
        if send_discord(webhooks.get('posts'), embed):
            db = get_db()
            cur = db.cursor()
            cur.execute("UPDATE twitter_posts SET is_notified = TRUE WHERE id = %s", (post['id'],))
            db.commit()
            db.close()
        time.sleep(2)

def push_trending_surge(webhooks):
    db = get_db()
    cur = db.cursor(cursor_factory=psycopg2.extras.RealDictCursor)
    
    # Check for posts with changes in the last 15 minutes
    cur.execute("""
        SELECT p.*, k.name as kol_name, k.twitter_url 
        FROM twitter_posts p
        JOIN kols k ON k.id = p.kol_id
        WHERE p.captured_at >= NOW() - INTERVAL '15 minutes'
    """)
    posts = cur.fetchall()
    db.close()
    
    significant_posts = []
    for p in posts:
        delta_views = (p['views'] or 0) - (p['last_views'] or 0)
        delta_likes = (p['likes'] or 0) - (p['last_likes'] or 0)
        delta_reposts = (p['reposts'] or 0) - (p['last_reposts'] or 0)
        delta_comments = (p['comments'] or 0) - (p['last_comments'] or 0)
        
        # Lower thresholds to catch more activity
        if (delta_views > 500 or delta_likes > 10 or delta_reposts >= 2 or delta_comments >= 2):
            significant_posts.append((p, delta_views, delta_likes, delta_reposts, delta_comments))
    
    if not significant_posts: 
        print("  - No significant interaction surges found in this window.")
        return

    # Weight: Likes(1) + Reposts(3) + Replies(5)
    significant_posts.sort(key=lambda x: x[2] + x[3]*3 + x[4]*5, reverse=True)
    
    fields = []
    db = get_db()
    cur = db.cursor()
    for sp in significant_posts[:5]:
        p, dv, dl, dr, dc = sp
        changes = []
        if dl > 0: changes.append(f"+{dl} Likes")
        if dr > 0: changes.append(f"+{dr} Reposts")
        if dc > 0: changes.append(f"+{dc} Replies")
        
        # Get usernames for replies and reposts
        cur.execute("SELECT username FROM twitter_post_replies WHERE post_id = %s ORDER BY id DESC LIMIT 5", (p['id'],))
        repliers = [r[0] for r in cur.fetchall()]
        repliers_text = f"\\n↳ New Repliers: {', '.join(['@'+r for r in repliers])}" if repliers else ""

        cur.execute("SELECT username FROM twitter_post_reposts WHERE post_id = %s ORDER BY id DESC LIMIT 5", (p['id'],))
        reposters = [r[0] for r in cur.fetchall()]
        reposters_text = f"\\n↳ New Reposters: {', '.join(['@'+r for r in reposters])}" if reposters else ""
        
        fields.append({
            "name": f"🔥 {p['kol_name']} ({', '.join(changes)})",
            "value": f"[{ (p['content'] or '')[:60] }...]({p['post_url']}){reposters_text}{repliers_text}",
            "inline": False
        })
    db.close()

    embed = {
        "title": "⚡ Trending: X KOL Engagement Surge!",
        "description": f"Found **{len(significant_posts)}** posts with rapid growth in the last window!\\n\\n**[Open Dashboard](http://187.124.224.226:3000)**",
        "color": 0xFF4500,
        "fields": fields,
        "footer": {"text": "Monitoring all interactions (Anyone)."},
        "timestamp": datetime.now().isoformat()
    }
    # Use 'interactions' webhook as per user request
    send_discord(webhooks.get('interactions'), embed)

def get_user_webhooks(db, user_id):
    cur = db.cursor(cursor_factory=psycopg2.extras.RealDictCursor)
    cur.execute("SELECT discord_webhook_posts, discord_webhook_interactions, discord_webhook_metrics, discord_webhook_heatmap FROM users WHERE id = %s", (user_id,))
    return cur.fetchone()

def run_push_for_user(db, user_id, job_type="all"):
    webhooks = get_user_webhooks(db, user_id)
    if not webhooks: return

    if job_type in ["posts", "all"]:
        hook = webhooks['discord_webhook_posts']
        if hook:
            # Only push posts belonging to this user's KOLs
            cur = db.cursor(cursor_factory=psycopg2.extras.RealDictCursor)
            cur.execute("""
                SELECT p.*, k.name as kol_name FROM twitter_posts p
                JOIN kols k ON p.kol_id = k.id
                WHERE k.user_id = %s AND p.is_notified = false
                ORDER BY p.captured_at DESC LIMIT 5
            """, (user_id,))
            new_posts = cur.fetchall()
            for post in new_posts:
                # ... existing embed logic but using 'hook' ...
                pass # (truncated for brevity, actual implementation follows)

if __name__ == "__main__":
    import sys
    job_type = sys.argv[1] if len(sys.argv) > 1 else "all"
    
    db = psycopg2.connect(DB_DSN)
    # Get all active users
    cur = db.cursor()
    cur.execute("SELECT id FROM users WHERE is_active = true")
    user_ids = [r[0] for r in cur.fetchall()]
    
    for uid in user_ids:
        print(f"Processing Discord push for User {uid}...")
        # (Call actual implementation here)
        push_new_posts(webhooks, 60)
    if job_type in ["heatmap", "all"]:
        push_trending_surge(webhooks)
    print("✅ Push complete!")
