requirements.txt
fastapi
uvicorn
psycopg2-binary
apscheduler
playwright
